refactor: reorganize and split `int_unique` module
