+++ crates/huub/src/solver/solving_context.rs
@@ -372,6 +372,32 @@ impl<'a> SolvingContext<'a> {
 		debug_assert_eq!(_prev, None);
 	}
 
+	/// Pop the next propagator from the queue and run it exactly once,
+	/// returning its propagation result. Unlike [`Self::run_propagators`],
+	/// this never advances past a single propagator, so the caller can inspect
+	/// the effect of one propagator in isolation (the literals it propagated
+	/// are left in `state.propagation_queue`).
+	///
+	/// Panics if the propagator queue is empty.
+	#[cfg(test)]
+	pub(crate) fn run_next_propagator(
+		&mut self,
+		propagators: &mut [BoxedPropagator],
+	) -> Result<(), Conflict<Decision<bool>>> {
+		let p = self
+			.state
+			.propagator_queue
+			.pop()
+			.expect("`run_next_propagator` called with an empty propagator queue");
+		self.current_prop = PropRef::from_raw(p);
+		let res = propagators[self.current_prop.index()]
+			.as_mut()
+			.propagate(self);
+		self.state.statistics.propagations += 1;
+		self.current_prop = PropRef::INVALID;
+		res
+	}
+
 	/// Run the propagators in the queue until a propagator detects a conflict,
 	/// returns literals to be propagated by the SAT solver, or the queue is
 	/// empty.
@@ -402,32 +428,6 @@ impl<'a> SolvingContext<'a> {
 				return;
 			}
 		}
-	}
-
-	/// Pop the next propagator from the queue and run it exactly once,
-	/// returning its propagation result. Unlike [`Self::run_propagators`],
-	/// this never advances past a single propagator, so the caller can inspect
-	/// the effect of one propagator in isolation (the literals it propagated
-	/// are left in `state.propagation_queue`).
-	///
-	/// Panics if the propagator queue is empty.
-	#[cfg(test)]
-	pub(crate) fn run_next_propagator(
-		&mut self,
-		propagators: &mut [BoxedPropagator],
-	) -> Result<(), Conflict<Decision<bool>>> {
-		let p = self
-			.state
-			.propagator_queue
-			.pop()
-			.expect("`run_next_propagator` called with an empty propagator queue");
-		self.current_prop = PropRef::from_raw(p);
-		let res = propagators[self.current_prop.index()]
-			.as_mut()
-			.propagate(self);
-		self.state.statistics.propagations += 1;
-		self.current_prop = PropRef::INVALID;
-		res
 	}
 }
